Add WelcomePage, Layout, Header, and Footer components; update dependencies
